[Simulator] Fix startup when using a new file for radio data (file is now properly created, errors from Storage are ignored); Do not start simulator at all in case of data errors; Fix startup option dialog when finding radio type from profile for first time; Fix RadioOptions struct version control; Cleanup unused WinConsole code; Fix a warning and a missing const reference; Move ~SimulatedUIWidget() to public.

+++ companion/src/simulation/widgets/radioswitchwidget.h
@@ -104,7 +104,7 @@ class RadioSwitchWidget : public RadioWidget
 
     void setToggleLocked(bool lock)
     {
-      if (m_flags != lock) {
+      if (m_flags != (quint16)lock) {
         m_flags = lock;
         setValue((int)lock);
         emit flagsChanged(m_flags);
